requirements.txt
# Core dependencies
langchain
langchain-community
langchain-openai
python-dotenv
huggingface_hub==0.23.2

# Web search and scraping
requests
beautifulsoup4
ddgs
wikipedia-api
wikipedia
feedparser  # For ArXiv API

# Data handling
pydantic==2.10.6
dataclasses-json

# Optional (for enhanced features)
urllib3
lxml  # Better HTML parsing

# Visualization
matplotlib

# Gradio
gradio==4.44.1
gradio_client==1.3.0